+++ dart/geobase/lib/src/codes/coords.dart
@@ -4,108 +4,46 @@
 //
 // Docs: https://github.com/navibyte/geospatial
 
-/// An enum for coordinate types.
+/// An enum for coordinate types (by spatial dimension and whether is measured).
 enum Coords {
-  /// Projected or cartesian coordinates (x, y).
-  ///
-  /// Coordinates are 2D and not measured.
+  /// 2D coordinates as projected (x, y) or geographic (lon, lat).
   xy(
     coordinateDimension: 2,
     spatialDimension: 2,
     is3D: false,
     isMeasured: false,
-    isGeographic: false,
     idWkb: 0,
     specifierWkt: null,
   ),
 
-  /// Projected or cartesian coordinates (x, y, z).
-  ///
-  /// Coordinates are 3D and not measured.
+  /// 3D coordinates as projected (x, y, z) or geographic (lon, lat, elev).
   xyz(
     coordinateDimension: 3,
     spatialDimension: 3,
     is3D: true,
     isMeasured: false,
-    isGeographic: false,
     idWkb: 1000,
     specifierWkt: 'Z',
   ),
 
-  /// Projected or cartesian coordinates (x, y, m).
-  ///
-  /// Coordinates are 2D and measured.
+  /// 2D measured coordinates as projected (x, y, m) or geographic (lon, lat,
+  /// m).
   xym(
     coordinateDimension: 3,
     spatialDimension: 2,
     is3D: false,
     isMeasured: true,
-    isGeographic: false,
     idWkb: 2000,
     specifierWkt: 'M',
   ),
 
-  /// Projected or cartesian coordinates (x, y, z, m).
-  ///
-  /// Coordinates are 3D and measured.
+  /// 3D measured coordinates as projected (x, y, z, m) or geographic (lon, lat,
+  /// elev, m).
   xyzm(
     coordinateDimension: 4,
     spatialDimension: 3,
     is3D: true,
     isMeasured: true,
-    isGeographic: false,
-    idWkb: 3000,
-    specifierWkt: 'ZM',
-  ),
-
-  /// Geographic coordinates (longitude, latitude).
-  ///
-  /// Coordinates are 2D and not measured.
-  lonLat(
-    coordinateDimension: 2,
-    spatialDimension: 2,
-    is3D: false,
-    isMeasured: false,
-    isGeographic: true,
-    idWkb: 0,
-    specifierWkt: null,
-  ),
-
-  /// Geographic coordinates (longitude, latitude, elevation).
-  ///
-  /// Coordinates are 3D and not measured.
-  lonLatElev(
-    coordinateDimension: 3,
-    spatialDimension: 3,
-    is3D: true,
-    isMeasured: false,
-    isGeographic: true,
-    idWkb: 1000,
-    specifierWkt: 'Z',
-  ),
-
-  /// Geographic coordinates (longitude, latitude, m).
-  ///
-  /// Coordinates are 2D and measured.
-  lonLatM(
-    coordinateDimension: 3,
-    spatialDimension: 2,
-    is3D: false,
-    isMeasured: true,
-    isGeographic: true,
-    idWkb: 2000,
-    specifierWkt: 'M',
-  ),
-
-  /// Geographic coordinates (longitude, latitude, elevation, m).
-  ///
-  /// Coordinates are 3D and measured.
-  lonLatElevM(
-    coordinateDimension: 4,
-    spatialDimension: 3,
-    is3D: true,
-    isMeasured: true,
-    isGeographic: true,
     idWkb: 3000,
     specifierWkt: 'ZM',
   );
@@ -116,7 +54,6 @@ enum Coords {
     required this.spatialDimension,
     required this.is3D,
     required this.isMeasured,
-    required this.isGeographic,
     required this.idWkb,
     required this.specifierWkt,
   });
@@ -133,44 +70,30 @@ enum Coords {
   /// True if coordinates represents a measured position with m coordinate.
   final bool isMeasured;
 
-  /// True for geographic coordinates (with longitude and latitude).
-  ///
-  /// If false, then coordinates are projected or cartesian (with x and y).
-  final bool isGeographic;
-
   /// The WKB type for coordinates, ie. `1000` for coordinates with Z.
-  /// 
+  ///
   /// Expected values are:
   /// * `0` for 2D coordinates: (x,y) or (lon,lat)
   /// * `1000` for 3D coordinates: (x,y,z) or (lon,lat,elev)
   /// * `2000` for measured coordinates: (x,y,m) or (lon,lat,m)
   /// * `3000` for 3D / measured coordinates: (x,y,z,m) or (lon,lat,elev,m)
-  /// 
-  /// References: 
+  ///
+  /// References:
   /// * [Simple Feature Access - Part 1: Common Architecture](https://www.ogc.org/standards/sfa)
   final int idWkb;
 
   /// An optional WKT specifier for coordinates, ie. `Z`, `M` or `ZM`.
   final String? specifierWkt;
 
-  /// Selects a [Coords] enum based on [isGeographic], [is3D] and [isMeasured].
+  /// Selects a [Coords] enum based on [is3D] and [isMeasured].
   static Coords select({
-    required bool isGeographic,
     required bool is3D,
     required bool isMeasured,
   }) {
-    if (isGeographic) {
-      if (is3D) {
-        return isMeasured ? Coords.lonLatElevM : Coords.lonLatElev;
-      } else {
-        return isMeasured ? Coords.lonLatM : Coords.lonLat;
-      }
+    if (is3D) {
+      return isMeasured ? Coords.xyzm : Coords.xyz;
     } else {
-      if (is3D) {
-        return isMeasured ? Coords.xyzm : Coords.xyz;
-      } else {
-        return isMeasured ? Coords.xym : Coords.xy;
-      }
+      return isMeasured ? Coords.xym : Coords.xy;
     }
   }
 }